Add files via upload
Now the student academic history place is completed. It can not only display all the courses that have been added to the course taken list and remove any courses from the course taken list, but also respond to admin's manipulation like the update of the course information and the removal of courses from the course list.